requirements.txt
kivy==2.2.1
kivymd==1.1.1
pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
pymysql==1.1.0
flask==2.3.3
flask-sqlalchemy==3.0.5 